services/orchestrator/pyproject.toml
[tool.poetry]
name = "personal-ai-assistant-orchestrator"
version = "0.1.0"
description = "Backend orchestrator service for the Personal AI Assistant"
authors = ["You <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.29"
aiosqlite = "^0.20.0"
openai = "^1.70.0"
anthropic = "^0.49.0"
httpx = "^0.27.0"
beautifulsoup4 = "^4.12.3"
duckduckgo-search = "^8.0.0"
ddgs = "^9.5.0"
pypdf = "^4.2.0"
python-docx = "^1.1.2"
python-pptx = "^0.6.23"
apscheduler = "^3.10"
argon2-cffi = "^25.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
stripe = "^11.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"

[tool.poetry.group.voice]
optional = true

[tool.poetry.group.voice.dependencies]
numpy = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
